REL-2377 Move PITLauncher to its own bundle #time 1h
