+++ bundle/edu.gemini.pit.launcher/build.sbt
@@ -1,0 +1,19 @@
+import OcsKeys._
+
+// note: inter-project dependencies are declared at the top, in projects.sbt
+
+name := "edu.gemini.pit.launcher"
+
+version := pitVersion.value.toOsgiVersion
+
+osgiSettings 
+
+ocsBundleSettings
+
+OsgiKeys.bundleActivator := None
+
+OsgiKeys.bundleSymbolicName := name.value
+
+OsgiKeys.dynamicImportPackage := Seq("")
+
+OsgiKeys.exportPackage := Seq()
